Remove tablespace support
WAL-E in Heroku production relies on obscure bug to work, namely,
broken symbolic links relative to the execution environment of WAL-E.

In more detail:

In Heroku, the only tablespaces used are those for temporary table
spaces, implemented with local (not EBS) disks seen in Amazon.  These
appear at `/mnt/tmp/$DIRECTORY` in the host file system, and `/tmp`
within the container running Postgres.  The `ephemeral` tablespace,
used for temporary assets, is symlinked into a subdirectory of `/tmp`.

This same link appears "broken" on the host: the path
`/tmp/pg_tblspc/...` does not exist outside the Postgres container.
This causes the following conditional in WAL-E to back up the symlink
rather than dereferencing it:

    if os.path.islink(ts_path) and os.path.isdir(ts_path):

When `ts_path` is a broken link, `os.path.isdir` returns false, and
the result of the tablespace detection code is not run.

Later on, I have been changing code such that WAL-E is run in a file
system layout with parity to the container, hence the `pg_tblspc`
symlink is no longer broken.  The result was restores would no longer
work.

To fix this, disable the tablespace support entirely, but only as a
Heroku-specific hack: in principle a `--no-tablespace-dereference`
flag (or similar) could be added to the `backup-push` step, but this
is so obscure that, for now, it does not seem to be worth the bloat to
the project at large.